Open Merge Request via git send-email
`git send-email` is a built-in `git` tool that allows users to send
patches via email. You can currently open a merge request by sending an
email to a specific email address, but it requires the patches be
attachments. `git send-email` sends the patch as the email body, so some
changes need to be made to allow for merge requests to be processed if
they are sent this way.