Allow setting Gitaly feature flags for individual projects
When we call Gitaly and set feature flags, we can't easily guarantee
that we can get the current project (if there even is such a thing).
However, there are some cases where we do have easy access to a project,
if one exists: when we're performing Git HTTP or SSH operations.

This allows - but doesn't require - Gitaly feature flags to be set for a
project, as long as the call site supports it. In this commit, that
means the two cases above (Git HTTP and SSH).