Fix unencoded image diff URLs
Some image diff URLs have characters like "#"
in them, which breaks the loader and causes
the diff viewer to just show the broken image
indicator. This is obviously bad.

The changes here simply escape spaces and
octothorpes ("hashtags") into their HTML-safe
percent-encoded ASCII values.

Notably, we CAN'T use something like
encodeURIComponent, because that would also
encode `/`, and this is actually used to load
the image, so that would be broken.