Update column 'updated_at' in container after repository storage move
We have several scripts in place that replicate some statistics
information to other databases. Some of them depend on the updated_at
column to identify the models they need to extract.

If we don't update the `updated_at` of the container after a
repository storage move, the scripts won't know that they need to
sync them.

See https://gitlab.com/gitlab-data/analytics/-/issues/7868