Add post migration to backfill projects updated at after repository move
In https://gitlab.com/gitlab-data/analytics/-/issues/7868 we discovered
that, after a repository was moved to a different storage, the
container's updated_at column of that repository wasn't updated.

This made some syncing scripts to Sisense to fail because they
couldn't detect the containers were updated.

In this commit, we add a post_migration to update those projects'
updated_at column that hasn't been updated after the repository was
moved.