WIP: Use previous license term for seat overage
To get the current seat overage the historical max data is pulled for
the past year. But not every license runs for a year and there was also
an overlap for new licenses to the old one. The overlap was removed with
customers-gitlab-com!2635. This change will take a previous license's
term to pull the historical max data and only use the past year as a
fallback.