Use previous license term for seat overage
To get the current seat overage the historical max data is pulled for
the past year. But not every license runs for a year. There was also an
overlap for new licenses to the old one which was removed with
https://gitlab.com/gitlab-org/customers-gitlab-com/-/merge_requests/2635

This change will take a previous license's term to pull the historical
max data and only use the past year as a fallback.