Track elasticsearch_timed_out_count for all ES requests in log
Since we've recently introduced timeouts in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/53435 we need some
way to track this in logs. The reason it needs special handling is that
Elasticsearch timeouts by default will just return early and return
partial results. For our use case this is what we want as the user still
gets something but we don't starve Elasticsearch resources but this
means there is no error propagated in the stack or the logs for us to
measure so we need to explicitly log this detail that is returned from
Elasticsearch.