Make the Geo OAuth application trusted by default
When a Geo secondary node is set up, we create a System
OAuth application, which is currently untrusted by
default.

This means every user that tries to login on a new
secondary will first receive this authorization popup.

Considering that Geo is actually "provided by GitLab",
the popup seems unnecessary, and we could set the
application as trusted by default.