Detect missing bucket in consolidated object storage settings
If you omit a single line from being defined and don't selectively
disable it, the previous implementation would allow a `nil` bucket
instead of raising an error. For example, in Omnibus, if you did the
following:

```
gitlab_rails['object_store']['enabled'] = true
gitlab_rails['object_store']['objects']['uploads']['bucket'] = {}
```

The parser would allow this configuration because it assumed the object
type wasn't activated, but by default it is if consolidated settings are
enabled.