Drop IP address input when creating audit events
To avoid issue of injecting a wrong IP address, we are no longer
accepting this information when invoking the `AuditEventService`. One
incorrect example is `current_sign_in_ip` which could be sometimes stale
if the user has not logged out. `AuditEventService` is capable of
resolving IP address directly from the request context.