Repository removal uses the Gitaly RPC
Gitaly has multiple ways to remove data from disk, using an `rm -rf` on
a directory (NamespaceService), or better the RemoveRepository RPC. The
latter RPC has logic that prevents stale data. In essence what the
ShellDestroyService and GitLabShellWorker used too. Given it's duplicate
this change removes the logic in favour of the logic at Gitaly.

This means that repository removal can be queued as soon as the
repository can be removed, and no delays are required when scheduling.