Don't move repositories before deleting
When a repository is deleted it used to be moved to
`<path>+deleted.git`, which made it inaccessable for new RPC requests.
That way the repositories were actually removed, even on networked
filesystems that aren't fully POSIX compliant.

This change removes most cases of the moves before deletions, as Gitaly
does that with the RemoveRepository RPC for us.