Migrate header part of Merge Conflicts Show App to Vue SFC
We just are interested in moving the Template layer for now. Simplest
way to do that iteratively, is to keep the original Vue Component in
place and create a new SFC, which directly references "$root". This
allows us to migrate bit by bit and we can deal with the data layer
later.