Improved award emoji picker
This improves the UX and performance of the award
emoji picker.

Groups the emojis into 'tabs' that are just anchors to the groups.

Renders the list of emoji categories with a virtual
scroller that only renders what is visible.
With this we also only render the visible categories emojis
instead of rendering all emojis.
Doing this reduces the DOM node count a lot.

[ci skip]