Unstick perf bar and move under navbar
Currently any sticky or full-height things need alternate
variations depending on if the perf bar is open or not.
This makes CSS messier and often causes slight but annoying
layout bugs. There is no particular need for the perf bar to be
fixed position (that I am aware of).

Doesn't move the one in fullscreen template (Web IDE only??)
as that was already static position.