Add mutation to accept merge requests
This is based mainly on the merge requests controller, since that
supports a wider range of merge strategies.

This verifies that we can merge:

- immediately
- MWPS
- using merge trains
- adding to merge trains when pipeline succeeds

Includes:

Move note mutation descriptions to classes

This is a rubocop appeasement change that moves the descriptions to the
mutation definitions, keeping the mutation mount points cleaner.