Add "New Repo" experiment to test familiarity
This experiment is testing the hypothesis that people coming to
GitLab from other products like GitHub and Bitbucket are familiar
with the "repository" term and are looking for it in our navigation.

By not having it, we are causing confusion and cognitive load for
users which may ultimately cause them to fail to adopt Create with
GitLab.

* adds gitlab_experiment.js which mimics the gitlab-experiment gem
  "run" implementation

part of:
https://gitlab.com/gitlab-org/gitlab/-/issues/285286