Update to ApolloClient@3
This is an attempt to update to ApolloClient@3. Multiple packages have
been replaced by a single one. This will likely result in a red pipeline
because Caches are now immutable. But it should be a good starting point
to explore what goes wrong:

TODOs:
1. Check that caching works correctly
2. Check that startup graphql works correctly
3. Update documentation with correct links
4. Link against importing from `@apollo/client`, we should always import
from `@apollo/client/core` as we otherwise pull in React related
dependencies.