Support automatic transitions of Jira issues
Add a new setting to the Jira integration to attempt an "automatic"
transition of Jira issues, by checking the available workflow
transitions on the issue and using the first one with a status
category of "Done".

To make this behaviour optional we also introduce an explicit checkbox
to disable issue transitions altogether.