Use learn gitlab utlimate template for experiment
In the learn gitlab experiment we want to link to issues that are only
available in the learn-gitlab-ultimate project template, since the
actions require a trial.
With this MR, we use the ultimate template when the experiment is active
and also update the issue IDs to link to the right issue to get more
information.