Collapse Deployments in Merge Request if Many
If there are more than 4 deployments an MR has been deployed to (either
pre- or post-merge), put them in a collapsible container and collapse it
by default.

This does require a bit of expansion on the MrCollapsibleContainer
slightly, providing a slot to add "header text", which is not blue nor
clickable. This was done without touching the current slot or props, and
should be completely backwards compatible.