Use full lifecycle in resolve_field
This makes resolve_field and resolve more complex so that they do not
simulate the execution of a field, but run it precisely as it is run in
the framework.

Doing this makes refactoring much safer, as we can have assurance that
our tests are in line with the real execution environment.