Include MRs for merge commits for changelogs
When generating a changelog, we now include links to merge requests for
any merge commits to use as input. Prior to this commit we'd only
consider merge requests that contained a commit.

The process of fetching merge requests by their merge SHA is a bit
different (and easier) compared to fetching them by their diffs. In this
commit we do so using an extra database query, but this query is simple
and should not pose any problems.

This fixes https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1579