Reduce queries when ticking runner queue
When UpdateBuildQueueService runs it iterates already
through the available runners for a given project.
In order to `tick_runner_queue` for each runner we
checked whether the `can_pick?` the build which
was checking again whether the runner was in the list
of available runner for the given project.

This change removes this redundant query which is very
expensive given that it's a UNION of 3 expensive
subqueries.