Add environment scope to group CI variables
Inspired by project level scoped variables, store an environment
scope for each group variable and use existing matching logic to
select only those that are relevant when passing variables to a
build.

Because the default scope is '*' (all environments) and there is
not yet a way to modify the scope, these changes alone don't
cause a change in behaviour.

https://gitlab.com/gitlab-org/gitlab/-/merge_requests/55256