Document Elasticsearch migrations must be finished before major upgrade
This is the first step to making it clear to operators about what they
need to check before proceeding with a major version upgrade as well as
hopefully helping them debug and fix any problems before proceeding.

We are also going to be removing the migration code from previous
versions in the major version upgrade so it's very important the
migration completes before the upgrade otherwise it will just fail as
it will be a blank migration.