Add Elasticsearch integration docs to recreate index from scratch
This documentation will likely help customers resolve a range of
problems. We regularly see customers trying to solve index problems by
guessing at a variety of different rake tasks or buttons in the UI. This
is the only sure way to fix a broken index state and it could hopefully
help them from running a bunch of other steps that are designed for
different purposes.

This will also be necessary for anyone that accidentally skips the
documentation for major version upgrades and misses some migrations
before they are deleted.