BulkImports: Add fields to track concurrent work
To enable BulkImports to run in smaller and more concurrent jobs, it's
needed to track the status, for each pipeline. Each pipeline will be
grouped in a stage, pipelines from the same stage will be able to run
in parallel.

The concurrency of the Pipelines will be added in another commit, for
better iteration. The main goal of this commit is to add the required
database fields and refactor the code accordingly to support the needs
of the next step.