Geo Node Status 2.0 - Primary Details
This change replaces the placeholder text
for the Primary specific node details
with actual components.

This change also implements the progress
bar component that is used here and in
will be used more in future parts
of this UI.

This change still leaves placeholder text in the secondary
node details section.