Allow retries for Advanced Search reindex operations
This change uses manual slicing when reindexing using
the Elasticsearch Reindex API. The number of slices
created is equal to Two times the number
of shards for each index. Slices are given a task id
in Elasticsearch and retried if the slice fails or
the total documents does not equal the number of
documents created, updated, and deleted. Total count
of documents between each new and old index is still
verified. A retry maximum is set so that reindexing
will fail if the retry limit is reached.