Move scanner UI strings to frontend
This moves the name, description and help paths for each of the scanners
in the Security Configuration page from the backend to the frontend.
This improves consistency between the Ultimate and non-Ultimate versions
of the page, and is one step along the path of sharing more code between
them (see https://gitlab.com/groups/gitlab-org/-/epics/5460 for more
details).

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/323376.