Encapsulate Deployment List in Own Component
In 18da847f4, the ability to collapse deployments on a merge request if
there were many was added. This made the pipeline component grow
slightly too large in its responsibilities.

The extra behaviour has been moved to the new deployments list
component. I think more might be able to be moved, but I wanted to avoid
taking in the whole merge request as a `prop`. I would be ameniable to
moving more props lower though.